x86_64: port kernel arch, IPC ABI, drivers, servers, and userland to amd64
Complete x86_64 MINIX port on the dev-efi3 branch. The kernel now builds as
a 64-bit ELF linked at 0xFFFFFFFF80400000 and boots via UEFI multiboot2
EFI64 entry.

Kernel / arch:
- head.S: UEFI EFI64-only entry, 4-level boot page tables, RIP-relative
  addressing throughout; drop multiboot1 legacy path
- mpx.S, klib.S: AMD64 SysV ABI (rdi/rsi/rdx args, rax return); pushaq/popaq
  macros; iretq; fxsave/fxrstor (clang 3.6 compatible)
- io_in*.S, io_out*.S: rewritten for x86_64 ABI
- debugreg.S: ld_dr*/st_dr* using movq %rdi/%rax
- pg_utils.c: 4-level paging (PML4/PDPT/PD/PT), vir_bytes kern_vir_start
- memory.c: vm_memset, virtual_copy_f, data_copy*, arch_proc_init,
  arch_phys_map, arch_enable_paging; #if CONFIG_OXPCIE throughout
- protect.c, acpi.c, arch_system.c: LP64 type/cast fixes (uintptr_t),
  rsp0 rename, X86_FLAGS_USER
- Makefile.inc: -mcmodel=kernel -mno-red-zone; OBJS.kernel shadow objects
  for libexec/libsys/libminc to avoid R_X86_64_32 overflow from pre-built
  archives compiled with -mcmodel=small
- multiboot2.h: created sys/arch/amd64/include/multiboot2.h; removed
  minix/include/arch/x86_64/include/multiboot.h (x86_64 is multiboot2-only)
- kernel.lds: high-half layout at 0xFFFFFFFF80400000

IPC ABI:
- 96-byte messages on x86_64 (_MSG_PAYLOAD_SIZE=88); i386 unchanged (64-byte)
- SYSCALL + INT33 dual entry; per-CPU kernel stacks; usermapped IPC vectors
- com.h, ipc.h, syslib.h, ipcconst.h: 64-bit endpoint/phys fields

Drivers / servers / userland:
- drivers: bus/storage/net/hid/power/ramdisk Makefiles extended to x86_64;
  pci.c cast, rs232.c, memory.c driver guards
- servers: PM/MIB/IS/VFS signal frame, ELF types, procfs format strings
- libc arch/x86_64: _cpuid, getprocessor, read_tsc, sigreturn, ipc stubs,
  ucontext, brksize — all rewritten for AMD64 ABI
- libsys: stacktrace, clock_time, tsc_util, vm_cache LP64 fixes
- libexec, libminixfs, libnetdriver, libsffs, libvtreefs, libusb: type/fmt
- libkvm: PG_* aliases for amd64 PTEs
- sys/arch/amd64: asm.h IMPORT, cpu.h X86_FLAGS_USER, fpu.h, signal.h,
  frame.h, types.h register_t, param.h, mcontext gregs layout
- efiboot/bootx64: KLINK_MACHINE=amd64, reloc_x86_64, ELF linker script
- share/mk/bsd.own.mk: MKZFS/__MINIX guard; MKPCI/MKACPI both archs
- Disable: ZFS/osnet, ENABLE_NLS (binutils), netstat, filemon, biosboot
- Add libc stubs: _lwp_setprivate, __msync13
- APIC: apic_asm.S pushaq/popaq/iretq, apic.c vir_bytes MMIO
- docs/: kernel-arch, kernel-build-fixes, ipc-abi, apic, drivers, servers,
  fs, vm, userland, efiboot, libminixfs port notes
- CLAUDE.md: updated with x86_64 porting patterns

+++ minix/commands/format/format.c
@@ -123,7 +123,7 @@ void format_track(int ffd, unsigned type, unsigned cyl, unsigned head)
 	if (lseek(ffd, track_pos, SEEK_SET) == -1) {
 		fprintf(stderr,
 		"format: seeking to cyl %u, head %u (pos %lld) failed: %s\n",
-			cyl, head, track_pos, strerror(errno));
+			cyl, head, (long long)track_pos, strerror(errno));
 		exit(1);
 	}
 
@@ -155,7 +155,7 @@ void verify_track(int vfd, unsigned type, unsigned cyl, unsigned head)
 	if (lseek(vfd, track_pos, SEEK_SET) == -1) {
 		fprintf(stderr,
 		"format: seeking to cyl %u, head %u (pos %lld) failed: %s\n",
-			cyl, head, track_pos, strerror(errno));
+			cyl, head, (long long)track_pos, strerror(errno));
 		exit(1);
 	}
 
@@ -168,7 +168,7 @@ void verify_track(int vfd, unsigned type, unsigned cyl, unsigned head)
 		if (lseek(vfd, track_pos, SEEK_SET) == -1) {
 			fprintf(stderr,
 	"format: seeking to cyl %u, head %u, sector %u (pos %lld) failed: %s\n",
-				cyl, head, sector, track_pos, strerror(errno));
+				cyl, head, sector, (long long)track_pos, strerror(errno));
 			exit(1);
 		}
 
@@ -176,7 +176,7 @@ void verify_track(int vfd, unsigned type, unsigned cyl, unsigned head)
 		case -1:
 			fprintf(stderr,
 		"format: bad sector at cyl %u, head %u, sector %u (pos %lld)\n",
-				cyl, head, sector, track_pos);
+				cyl, head, sector, (long long)track_pos);
 			bad_count++;
 			break;
 		case SECTOR_SIZE:
@@ -184,7 +184,7 @@ void verify_track(int vfd, unsigned type, unsigned cyl, unsigned head)
 			break;
 		default:
 			fprintf(stderr, "format: short read at pos %lld\n",
-				track_pos);
+				(long long)track_pos);
 			bad_count++;
 		}
 		track_pos+= SECTOR_SIZE;
